Split yelpfunctions into two files for modularity Clean up code on back end
