+++ server/helpers/yelpFunctions.js
@@ -1,4 +1,6 @@
 var yelp = require('./yelp');
+var key = require('../api/api_key');
+var coord = require('./coordinateHelpers');
 
 // create yelp client using Oauth
 var yelpClient = yelp.createClient({
@@ -9,41 +11,13 @@ var yelpClient = yelp.createClient({
   ssl: process.env.SSL || key.ssl
 });
 
-// yelp search parameter configuration
+// Yelp search parameter configuration defaults
 var yelpProperty = {
-  term: "food",           // Type of business (food, restaurants, bars, hotels, etc.)
-  limit: 10,              // Number of entries returned from each call
-  sort: 2,                // Sort mode: 0=Best matched (default), 1=Distance, 2=Highest Rated
+  term: "food",             // Type of business (food, restaurants, bars, hotels, etc.)
+  limit: 10,                // Number of entries returned from each call
+  sort: 2,                  // Sort mode: 0=Best matched (default), 1=Distance, 2=Highest Rated
   radius_filter: 5*1609.34  // Search radius: 1 mile = 1609.3 meters, 5 miles is good for rural areas
 };
-
-// define the to Radian function
-if (typeof(Number.prototype.toRad) === "undefined") {
-  Number.prototype.toRad = function() {
-    return this * Math.PI / 180;
-  }
-}
-
-// calculate the distance between 2 waypoints, given their latitudes and longitudes, return distance in miles
-function calcDistance(pt1, pt2) {
-  var R = 6371; // earth radius, in km
-  var lat1 = pt1.location.coordinate['latitude'];
-  var lon1 = pt1.location.coordinate['longitude'];
-  var lat2 = pt2.location.coordinate['latitude'];
-  var lon2 = pt2.location.coordinate['longitude'];
-
-
-  var dLat = (lat2 - lat1).toRad();
-  var dLon = (lon2 - lon1).toRad();
-  var lat1 = lat1.toRad();
-  var lat2 = lat2.toRad();
-
-  var a = Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-          Math.sin(dLon / 2) * Math.sin(dLon / 2) * Math.cos(lat1) * Math.cos(lat2);
-  var c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  var distance = R * c * 0.621371; // convert distance from km to miles
-  return distance;
-}
 
 function isAlreadyInArray(array, target) {
   for (var i = 0 ; i < array.length; i++) {
@@ -54,10 +28,10 @@ function isAlreadyInArray(array, target) {
   return false;
 }
 
-var commonFilter=["McDonald's", "Burger King", "Jack in the Box", "Carl's Junior", "StarBucks", "Subway",
+// check if a place is a common place to be filtered out
+var commonFilter = ["McDonald's", "Burger King", "Jack in the Box", "Carl's Junior", "StarBucks", "Subway",
 "Pizza Hut", "Del Taco", "Taco Bell", "Chick-fil-A", "Farm", "Truck", "In-N-Out"];
 
-// check if a place is a common place to be filtered out
 function isCommonPlace(businessEntry, commonFilter){
   for (var i = 0; i < commonFilter.length; i++) {
     if (businessEntry.name.indexOf(commonFilter[i]) > -1)
@@ -66,45 +40,11 @@ function isCommonPlace(businessEntry, commonFilter){
   return false;
 }
 
-// parse google coordinate into {latitude:..., longitude: ... } format
-function parseGoogleCoord(googleCoord) {
-  var latitude = parseFloat(googleCoord.match(/^.*,/)[0].replace(",", ""));
-  var longitude = parseFloat(googleCoord.match(/,.*$/)[0].replace(",", ""));
-  var obj = {
-    location: {
-      coordinate : {
-        latitude: latitude,
-        longitude: longitude
-      }
-    }
-  }
-  return obj;
-}
-
-// trim the google waypoint coordinate to take out start and end way point so no clustering at 2 ends.
-function trimGoogleCoord(googleCoords, distance) {
-  var trimmedCoords = [];
-  //Loop through array and only push the coordinates that are distanceBetweenQueries apart
-
-  if (googleCoords.length > 5) {
-    for (var i=0; i<googleCoords.length; i++) {
-      if (calcDistance(parseGoogleCoord(googleCoords[i]), parseGoogleCoord(googleCoords[0])) >= distance/20 &&
-        calcDistance(parseGoogleCoord(googleCoords[i]), parseGoogleCoord(googleCoords[googleCoords.length-1])) >= distance/20) {
-        trimmedCoords.push(googleCoords[i]);
-      }
-    }
-  } else {
-    trimmedCoords = googleCoords;
-  }
-
-  return trimmedCoords;
-}
-
 // function to use yelp API to get the top choices based on longitude and latitude
 module.exports.searchYelp = function (req, res, googleCoords, distance, callback) {
   //Counter variable which will keep track of how many Yelp calls have completed
   //A separate counter is needed due to the asynchronous nature of web requests
-  var trimmedCoords= trimGoogleCoord(googleCoords, distance);
+  var trimmedCoords = coord.trimGoogleCoord(googleCoords, distance);
   var counter = 0;
   // Array that stores all of the Yelp results from all calls to Yelp
   var yelpResults = [];
@@ -113,11 +53,11 @@ module.exports.searchYelp = function (req, res, googleCoords, distance, callback
   yelpProperty.term = req.body.optionFilter;           // Type of business (food, restaurants, bars, hotels, etc.)
 
   if (distance <= 20) {
-    yelpProperty.radius_filter = 0.8*1609.34 ;
+    yelpProperty.radius_filter = 0.8 * 1609.34 ;
   } else if (distance <= 40) {
-    yelpProperty.radius_filter = 2.5*1609.34;
+    yelpProperty.radius_filter = 2.5 * 1609.34;
   } else {
-    yelpProperty.radius_filter = 5*1609.34;
+    yelpProperty.radius_filter = 5 * 1609.34;
   }
 
   //Request yelp for each point along route that is returned by filterGoogle.js
@@ -198,7 +138,7 @@ module.exports.createTopResultsJSON = function(yelpResults, distance) {
 
   for (var m = 1; m < allBusinesses.length; m) {
     // if next waypoint less than total distance/20 mi away
-    if (calcDistance(evenSpreadResults[n], allBusinesses[m]) < (distance / 20)) {
+    if (coord.calcDistance(evenSpreadResults[n], allBusinesses[m]) < (distance / 20)) {
       // then skip
       m++;
     }
