+++ server/router.js
@@ -28,5 +28,4 @@ router.get('/*', function (req, res) {
   res.redirect('/');
 });
 
-
 module.exports = router;
